Binding entity: OneToMany (users -> features)

+++ src/main/java/com/company/WebBoot/service/UserService.java
@@ -4,7 +4,7 @@ import com.company.WebBoot.model.Users;
 import java.util.List;
 
 public interface UserService {
-    Users getById(Long id);
+    Users getById(Integer id);
     void save(Users user);
     void delete(Long id);
     List<Users> getAll();
